backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
alembic
pydantic
pydantic-settings
python-dotenv
